@@ -12,16 +12,14 @@
     "axios": "^1.6.2",
     "bootstrap": "^4.6.0",
     "classnames": "^2.3.1",
-    "formik": "^2.2.9",
+    "formik": "^2.4.5",
     "lodash.debounce": "^4.0.8",
     "normalizr": "^3.6.2",
     "ramda": "0.29.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-redux": "^8.1.3",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.0.4",
+    "react-router-dom": "^6.21.1",
     "reactstrap": "^8.9.0",
     "recompose": "^0.30.0",
     "redux": "^5.0.1",
@@ -29,19 +27,16 @@
   },
   "devDependencies": {
     "@pact-foundation/pact": "^12.1.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^14.0.0",
+    "@testing-library/jest-dom": "^6.1.6",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
     "@types/auth0-js": "^9.14.5",
     "@types/jest": "^29.0.0",
     "@types/lodash.debounce": "^4.0.6",
     "@types/node": "^16.0.0",
     "@types/ramda": "^0.29.1",
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-helmet": "^6.1.2",
-    "@types/react-router-bootstrap": "^0.24.5",
-    "@types/react-router-dom": "^5.1.8",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
     "@types/rosie": "^0.0.45",
     "@types/yup": "^0.29.13",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
