@@ -1,5 +1,5 @@
 import React from 'react'
-import { Route } from 'react-router-dom'
+import { Route, Routes } from 'react-router-dom'
 import { screen } from '@testing-library/react'
 
 import BookList from './BookList'
@@ -10,9 +10,14 @@ jest.mock('rest/fetch')
 describe('components', () => {
   describe('BookList', () => {
     it('renders correctly', async () => {
-      fullRender(<Route path="/books/:year" component={BookList} />, {
-        route: '/books/2017',
-      })
+      fullRender(
+        <Routes>
+          <Route path="/books/:year" element={<BookList />} />
+        </Routes>,
+        {
+          route: '/books/2017',
+        }
+      )
 
       await screen.findByText('history')
       await screen.findByText('Time To Murder And Create')
