@@ -1,11 +1,11 @@
 import React from 'react'
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 import CreateButton from 'components/create_button'
 
 const BookCreator = () => {
-  const history = useHistory()
-  return <CreateButton onClick={() => history.push('/books/new')} />
+  const navigate = useNavigate()
+  return <CreateButton onClick={() => navigate('/books/new')} />
 }
 
 export default BookCreator
