@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react'
 
-import { Redirect } from 'react-router-dom'
+import { Navigate } from 'react-router-dom'
 import { useAppDispatch, useAppSelector } from 'hooks'
 
 import { actions, authenticatedSelector } from 'state/login'
 
 const redirectIfLoggedIn = (authenticated: boolean) => {
   if (authenticated) {
-    return <Redirect to="/" />
+    return <Navigate replace to="/" />
   }
 }
 
