@@ -1,8 +1,7 @@
 import React from 'react'
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 import { Field, Form, Formik, FormikProps } from 'formik'
 
-import { LinkContainer } from 'react-router-bootstrap'
 import { Button, Container, Row, Col } from 'reactstrap'
 
 import ReactstrapSelect from './ReactstrapSelect'
@@ -26,7 +25,7 @@ const categories: Categories[] = [
 ]
 
 const NewBookWrapper = () => {
-  const history = useHistory()
+  const navigate = useNavigate()
   const dispatch = useAppDispatch()
   return (
     <Formik
@@ -36,7 +35,7 @@ const NewBookWrapper = () => {
       // eslint-disable-next-line @typescript-eslint/require-await
       onSubmit={async (values, { setStatus }) =>
         dispatch(
-          actions.create(values, history, (error) =>
+          actions.create(values, navigate, (error) =>
             setStatus({ submitError: error })
           )
         )
@@ -50,56 +49,61 @@ const NewBookWrapper = () => {
 const NewBook = ({
   status: { submitError },
   isSubmitting,
-}: FormikProps<BookForm>) => (
-  <section>
-    <Form>
-      <Container fluid>
-        <Row>
-          <Col>
-            {input({ name: 'title', placeholder: 'Catch-22' })}
-            {input({ name: 'year', type: 'number' })}
-            {input({
-              name: 'description',
-              type: 'textarea',
-              placeholder: "That's some catch, that Catch-22",
-              rows: '3',
-            })}
+}: FormikProps<BookForm>) => {
+  const navigate = useNavigate()
+  return (
+    <section>
+      <Form>
+        <Container fluid>
+          <Row>
+            <Col>
+              {input({ name: 'title', placeholder: 'Catch-22' })}
+              {input({ name: 'year', type: 'number' })}
+              {input({
+                name: 'description',
+                type: 'textarea',
+                placeholder: "That's some catch, that Catch-22",
+                rows: '3',
+              })}
 
-            <Stars />
+              <Stars />
 
-            <Field
-              name="category"
-              label="category"
-              id="category"
-              component={ReactstrapSelect}
-              inputprops={{
-                name: 'category',
-                id: 'category',
-                options: categories,
-              }}
-            />
+              <Field
+                name="category"
+                label="category"
+                id="category"
+                component={ReactstrapSelect}
+                inputprops={{
+                  name: 'category',
+                  id: 'category',
+                  options: categories,
+                }}
+              />
 
-            <section>
-              {submitError && <div>{submitError}</div>}
-              <LinkContainer to="/books">
-                <Button className="float-right " color="secondary">
+              <section>
+                {submitError && <div>{submitError}</div>}
+                <Button
+                  onClick={() => navigate('/books')}
+                  className="float-right "
+                  color="secondary"
+                >
                   Cancel
                 </Button>
-              </LinkContainer>
-              <Button
-                className="float-right mr-2"
-                color="primary"
-                type="submit"
-                disabled={isSubmitting}
-              >
-                Create
-              </Button>
-            </section>
-          </Col>
-        </Row>
-      </Container>
-    </Form>
-  </section>
-)
+                <Button
+                  className="float-right mr-2"
+                  color="primary"
+                  type="submit"
+                  disabled={isSubmitting}
+                >
+                  Create
+                </Button>
+              </section>
+            </Col>
+          </Row>
+        </Container>
+      </Form>
+    </section>
+  )
+}
 
 export default NewBookWrapper
