@@ -1,6 +1,7 @@
 import React from 'react'
 import { screen } from '@testing-library/react'
 
+import { Route, Routes } from 'react-router-dom'
 import BooksArea from './BooksArea'
 import { fullRender } from 'test'
 
@@ -9,7 +10,12 @@ jest.mock('rest/fetch')
 describe('components', () => {
   describe('BooksArea', () => {
     it('renders correctly', async () => {
-      fullRender(<BooksArea />, { route: '/books/2016' })
+      fullRender(
+        <Routes>
+          <Route path="/books/:year" element={<BooksArea />} />
+        </Routes>,
+        { route: '/books/2016' }
+      )
 
       await screen.findByText('Add Book')
       await screen.findByText('Time To Murder And Create')
