@@ -1,7 +1,5 @@
 import React, { useState } from 'react'
-import { Link } from 'react-router-dom'
-import { LinkContainer } from 'react-router-bootstrap'
-import { Helmet } from 'react-helmet'
+import { Link, useNavigate } from 'react-router-dom'
 
 import {
   Collapse,
@@ -24,16 +22,11 @@ const Login = () => (
 
 const Header = () => {
   const [isOpen, setIsOpen] = useState(false)
+  const navigate = useNavigate()
   return (
     <header className="mb-4">
-      <Helmet>
-        <meta charSet="utf-8" />
-        <title>Shelf</title>
-      </Helmet>
       <Navbar dark color="dark" expand="sm">
-        <LinkContainer to="/books">
-          <NavbarBrand>Shelf</NavbarBrand>
-        </LinkContainer>
+        <NavbarBrand onClick={() => navigate('/books')}>Shelf</NavbarBrand>
         <NavbarToggler onClick={() => setIsOpen(!isOpen)} />
         <Collapse isOpen={isOpen} navbar>
           <Nav className="ml-auto" navbar>
