@@ -1,26 +1,28 @@
 import React from 'react'
 
-import { Router } from 'react-router-dom'
+import { MemoryRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
-import { createMemoryHistory, History } from 'history'
 import { render } from '@testing-library/react'
 import store from 'state'
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 const fullRender = (
   children: React.ReactNode,
-  {
-    route = '/',
-    history = createMemoryHistory({ initialEntries: [route] }),
-  }: { route?: string; history?: History } = {}
+  { route = '/' }: { route?: string } = {}
 ) => {
   return {
     ...render(
       <Provider store={store}>
-        <Router history={history}>{children}</Router>
+        <MemoryRouter initialEntries={[route]}>{children}</MemoryRouter>
       </Provider>
     ),
-    history,
+    store,
+  }
+}
+
+export const fullRenderWithoutRouter = (children: React.ReactNode) => {
+  return {
+    ...render(<Provider store={store}>{children}</Provider>),
     store,
   }
 }
