@@ -1,6 +1,6 @@
-import fullRender from './full_render'
+import fullRender, { fullRenderWithoutRouter } from './full_render'
 import createProvider from './create_provider'
 
 import * as factories from './factories'
 
-export { fullRender, factories, createProvider }
+export { fullRender, fullRenderWithoutRouter, factories, createProvider }
